Make compatible with Sprockets 4.0.0
The `Sprockets::SassFunctions` module is gone now, and replaced by two
others, corresponding to the different SASS engines that might be used.

Also, `Sprockets::Index` doesn't exist anymore, but instead we might
be returned a `Sprockets::CachedEnvironment`.

Finally, `Sprockets::Asset#pathname` is gone, replaced with
`#filename`, but this returns a string whereas other parts of this gem
assume that they can call `read` on it, so we need to re-wrap it in a
`Pathname` instance.